+++ src/components/ids-popup-menu/ids-popup-menu.ts
@@ -97,6 +97,11 @@ export default class IdsPopupMenu extends Base {
     if (this.hasOpenEvents) {
       this.hide();
     }
+    // Clean up any Mutation observers
+    if (this.#mo) {
+      this.#mo.disconnect();
+      this.#mo = undefined;
+    }
   }
 
   /**
@@ -143,6 +148,7 @@ export default class IdsPopupMenu extends Base {
     // When the underlying Popup triggers its "show" event, pass the event to the Host element.
     this.offEvent('show');
     this.onEvent('show', this.container, (e: CustomEvent) => {
+      this.hideOtherMenus();
       if (!this.parentMenuItem) {
         this.triggerEvent('show', this, e);
       }
@@ -155,11 +161,34 @@ export default class IdsPopupMenu extends Base {
       if (!this.parentMenuItem) {
         this.triggerEvent('hide', this, e);
       }
+      this.#mo?.disconnect();
     });
 
     // Set up all the events specifically-related to the "trigger" type
     this.refreshTriggerEvents();
   }
+
+  /** Hide any older (enterprise) menus */
+  hideOtherMenus() {
+    const openMenu = (document.querySelector('.popupmenu.is-open') as HTMLElement);
+    openMenu?.classList.remove('is-open');
+    const openMenuElem = (document.querySelector('[data-popupmenu].is-open') as HTMLElement);
+    openMenuElem?.classList.remove('is-open');
+
+    // Use the is-open class to close it
+    this.#mo = new MutationObserver((mutations) => {
+      mutations.forEach((mutationRecord) => {
+        const target = (mutationRecord.target as HTMLElement);
+        if (target?.className === 'is-open') {
+          this.hide();
+        }
+      });
+    });
+    const body = document.querySelector('body');
+    if (body) this.#mo.observe(document.querySelector('body') as Node, { attributes: true, subtree: true });
+  }
+
+  #mo: MutationObserver | undefined = undefined;
 
   /**
    * Sets up the connection to the global keyboard handler
@@ -260,7 +289,9 @@ export default class IdsPopupMenu extends Base {
    * @returns {void}
    */
   show(): void {
-    if (this.popup?.visible) return;
+    if (this.popup?.visible) {
+      this.hide();
+    }
 
     // Trigger a veto-able `beforeshow` event.
     if (!this.triggerVetoableEvent('beforeshow')) {