+++ src/components/ids-wizard/ids-wizard.ts
@@ -305,13 +305,11 @@ export default class IdsWizard extends Base {
   }
 
   /**
-   * updates hrefURIs at select points
+   * Updates hrefURIs at select points
    * so we don't need to recalculate
    * when setting clickable or step number
    * again; also allows us to easily run
-   * calculations to use unique-but-meaningful
-   * links
-   *
+   * calculations to use unique-but-meaningful links
    * @private
    */
   updateHrefURIs() {
@@ -338,9 +336,7 @@ export default class IdsWizard extends Base {
   }
 
   /**
-   * retrieves a step marker element within
-   * a wizard's shadow DOM
-   *
+   * Retrieves a step marker element within a wizard's shadow DOM
    * @param {*} wizardEl source Wizard
    * @param {*} stepNumber step number
    * @returns {HTMLElement} the step element