+++ doc/CHANGELOG.md
@@ -11,6 +11,7 @@
 - `[Input/TriggerField]` Web component now displays as `inline`, similar to HTMLInputElement. ([#1157](https://github.com/infor-design/enterprise-wc/issues/1157))
 - `[Popup]` Unset text align coming from HTML attribute. ([#1200](https://github.com/infor-design/enterprise-wc/issues/1200))
 - `[PopupMenu]` Added scrollable behavior to `max-height`-enabled popup menus. ([#1205](https://github.com/infor-design/enterprise-wc/issues/1205))
+- `[PopupMenu]` Added code to hide enterprise context menus when web context menus open and vise versa. ([#1179](https://github.com/infor-design/enterprise-wc/issues/1179))
 
 ## 1.0.0-beta.8